Fix overflow issues in LocalCache, as per Guava issue 1761 ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=68007129

+++ guava/src/com/google/common/cache/LocalCache.java
@@ -2016,7 +2016,7 @@ class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<K, V> 
      * The weight of the live elements in this segment's region.
      */
     @GuardedBy("Segment.this")
-    int totalWeight;
+    long totalWeight;
 
     /**
      * Number of updates that alter the size of the table. This is used during bulk-read methods to
